+++ src/com/softserve/academy/graduation/EmployeeHandler.java
@@ -1,0 +1,126 @@
+package softserve.academy.graduation;
+
+import java.util.*;
+
+public class EmployeeHandler {
+    final static ConsolePrinter PRINTER = new ConsolePrinter();
+    final static InputValidator VALIDATOR = new InputValidator();
+
+    // TODO: change hashSet to LocalDate
+    public void addEmployee(
+            Map<Integer, Employee> employeeMap,
+            String newName,
+            int newSalary,
+            String newPosition,
+            HashMap<String, Integer> newDateOfBirth,
+            String newEmail) {
+        Set<Integer> idSet = employeeMap.keySet();
+        Integer newId = Collections.max(idSet) + 1;
+        PRINTER.printLine();
+        if (VALIDATOR.isEmailUnique(employeeMap, newEmail)) {
+            employeeMap.put(
+                    newId,
+                    new Employee(
+                            newName,
+                            newSalary,
+                            newPosition,
+                            newDateOfBirth.get("year"),
+                            newDateOfBirth.get("month"),
+                            newDateOfBirth.get("day"),
+                            newEmail
+                    )
+            );
+            System.out.printf("New record for employee %s (ID:%d) was created!\n", newName, newId);
+        } else {
+            System.out.println("Employee creation failed!");
+        }
+    }
+
+    // TODO: change hashSet to LocalDate
+    public void editEmployee(
+            Map<Integer, Employee> employeeMap,
+            Integer targetId,
+            String newName,
+            int newSalary,
+            String newPosition,
+            HashMap<String, Integer> newDateOfBirth,
+            String newEmail) {
+
+        PRINTER.printLine();
+
+        // a little workaround to avoid self-comparing of current email and new one if they are the same;
+        String oldEmail = employeeMap.get(targetId).getEmail();
+        employeeMap.get(targetId).setEmail("");
+
+        if (VALIDATOR.isEmailUnique(employeeMap, newEmail)) {
+            Employee employee = employeeMap.get(targetId);
+            employee.setName(newName);
+            employee.setSalary(newSalary);
+            employee.setPosition(newPosition);
+            employee.setDateOfBirth(
+                    newDateOfBirth.get("year"),
+                    newDateOfBirth.get("month"),
+                    newDateOfBirth.get("day")
+            );
+            employee.setEmail(newEmail);
+            employeeMap.put(targetId, employee);
+            System.out.printf("Changes for %s (ID:%d) saved!\n", newName, targetId);
+        } else {
+            // if the new email has duplicate among other employees, we give back the old email to our instance;
+            // this is done in case we have duplicate emails in the starting set of employees;
+            employeeMap.get(targetId).setEmail(oldEmail);
+            System.out.println("Employee changes discarded!");
+        }
+    }
+
+    public void sortEmployeesByName(Map<Integer, Employee> employeeMap) {
+        List<Map.Entry<Integer, Employee>> employeeList = new ArrayList<>(employeeMap.entrySet());
+        employeeList.sort(Comparator.comparing(e -> e.getValue().getName()));
+        employeeMap.clear();
+        for (Map.Entry<Integer, Employee> entry : employeeList) {
+            employeeMap.put(entry.getKey(), entry.getValue());
+        }
+    }
+
+    public void sortEmployeesByPosition(Map<Integer, Employee> employeeMap) {
+        List<Map.Entry<Integer, Employee>> employeeList = new ArrayList<>(employeeMap.entrySet());
+        employeeList.sort(Comparator.comparing((Map.Entry<Integer, Employee> e) ->
+                e.getValue().getPosition()).thenComparingInt(e ->
+                e.getValue().getSalary()));
+        employeeMap.clear();
+        for (Map.Entry<Integer, Employee> entry : employeeList) {
+            employeeMap.put(entry.getKey(), entry.getValue());
+        }
+    }
+
+    public void sortEmployeesById(Map<Integer, Employee> employeeMap) {
+        List<Map.Entry<Integer, Employee>> employeeList = new ArrayList<>(employeeMap.entrySet());
+        employeeList.sort(Comparator.comparingInt(Map.Entry::getKey));
+        employeeMap.clear();
+        for (Map.Entry<Integer, Employee> entry : employeeList) {
+            employeeMap.put(entry.getKey(), entry.getValue());
+        }
+    }
+
+    public List<Task> getEmployeeTasks(Map<Integer, Task> taskMap, Integer employeeId) {
+        List<Task> tasksAssigned = new ArrayList<>();
+        for (Task task : taskMap.values()) {
+            if (task.getStaff().contains(employeeId)) {
+                tasksAssigned.add(task);
+            }
+        }
+        return tasksAssigned;
+    }
+
+    public void removeEmployeeById(Map<Integer, Employee> employeeMap, Map<Integer, Task> taskMap, int employeeId) {
+        if (VALIDATOR.isIdValid(employeeMap, employeeId)) {
+            employeeMap.remove(employeeId);
+            System.out.printf("Employee (%d) removed.\n", employeeId);
+            for (Task task : taskMap.values()) {
+                task.getStaff().remove(employeeId);
+            }
+        }
+    }
+
+
+}
